report reorg
