+++ backend/src/app/routers/report.ts
@@ -1,8 +1,11 @@
 import {Router} from 'express';
 import * as reportController from '../controllers/report';
 
+import { adminAuthorization } from '../../lib/authorization/admin';
+
 export const router: Router = Router({mergeParams: true});
 
+router.use(adminAuthorization);
 router.get('/project/hours', reportController.generateReportProjectByHours);
 router.get('/project/cost', reportController.generateReportProjectByCost);
 router.get('/user/hours', reportController.generateReportUserByHours);
