+++ backend/src/app/serializers/report.ts
@@ -1,30 +1,17 @@
-import {getUniqueRowNames, transformReportForFrontend} from "../../lib/report"
+import {transformBudgetReport, transformReportForFrontend} from "../../lib/report"
 
-export const getReportProjectByHours = (report: object[]) => {
+export const getReportProjectByHours = (report) => {
     return transformReportForFrontend(report, 'userName', 'projectName', 'timeSpent');
 }
-export const getReportProjectByCost = (report: object[]) => {
+export const getReportProjectByCost = (report) => {
     return transformReportForFrontend(report, 'userName', 'projectName', 'cost');
 }
-export const getReportUserByHours = (report: object[]) => {
+export const getReportUserByHours = (report) => {
     return transformReportForFrontend(report, 'projectName', 'userName', 'timeSpent');
 }
-export const getReportUserByCost = (report: object[]) => {
+export const getReportUserByCost = (report) => {
     return transformReportForFrontend(report, 'projectName', 'userName', 'cost');
 }
-export const getBudgetReport = (report: object[]) => {
-    let rowNames = getUniqueRowNames(report, 'projectName');
-    let finalObject = {};
-    rowNames.forEach(row => finalObject[row] = (new Object()));
-    Object.keys(finalObject).forEach(key => {
-        report.forEach(individualResult => {
-            if (individualResult['projectName'] === key) {
-                finalObject[key]['actualCosts'] = individualResult['actualCosts'];
-                finalObject[key]['budgetCosts'] = individualResult['budgetCosts'];
-                finalObject[key]['overUnder'] = individualResult['overUnder'];
-            }
-        });
-    })
-    return finalObject;
+export const getBudgetReport = (report) => {
+        return transformBudgetReport(report);
 }
-
