+++ backend/src/lib/report.ts
@@ -30,3 +30,19 @@ export const calculateAndAddTotal = (finalObject: object) => {
         finalObject[row]['total'] = sum;
     });
 }
+
+export const transformBudgetReport = (report) => {
+    let rowNames = getUniqueRowNames(report, 'projectName');
+    let finalObject = {};
+    rowNames.forEach(row => finalObject[row] = (new Object()));
+    Object.keys(finalObject).forEach(key => {
+        report.forEach(individualResult => {
+            if (individualResult['projectName'] === key) {
+                finalObject[key]['actualCosts'] = individualResult['actualCosts'];
+                finalObject[key]['budgetCosts'] = individualResult['budgetCosts'];
+                finalObject[key]['overUnder'] = individualResult['overUnder'];
+            }
+        });
+    })
+    return finalObject;
+}
